+++ CHANGELOG.md
@@ -76,6 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - edges/curves.py cut_between_two_points -> trim
 - defines ordering of curve methods
 - npy -> np
+- curves.py: curve_intersections -> intersections
 
 #### CI
 - Avoid calling babylonjs method in CI to reduce CI time